requirements.txt
pylightning==0.0.7.3
python-bitcoinlib==0.10.2